dev-requirements.txt
pytest==4.6.2
pytest-cov==2.7.1
-r requirements.txt
pandas>=1.0.0
pandas_ods_reader
liquer-framework>=0.3.1
matplotlib
plotly==4.9.0
pygments
xlsxwriter